admin/requirements.txt
flask>=3.0.0
pillow>=10.0.0
python-dateutil>=2.8.0
